Add files via upload
updated form submission to send users to a thank you page after submitting info. The first 4 text boxes must be filled out before the user can hit the submit button. If the gate code box is checked but nothing is in the text box of the gate code the user must fill that in before submitting. In short, this update will ensure the user enters all the vital information before hitting submit. Now the webpage is ready to be upgraded to have the capability of sending the information the user enters to our professional business email.